암호화 복호화시 결과 파일이 존재하면 중단 및 메세지 뿌리기

+++ OneFileEncryptDecrypt/XWork/DecryptWork.cs
@@ -8,55 +8,68 @@ namespace OneFileEncryptDecrypt.XWork
     {
         public static void ExecuteNow(XAppSettings.AppSettingsX asx, XConsole.ConsoleWriteMessageSet cwms, XModel.CryptoWorkOrder cwo)
         {
-            // 저장 할 파일들 경로생성
-            var cfn = asx.Crypto.GetCryptoWorkPath;
-            // 복호화 키 생성
-            var cryptoKey = XCrypto.AES256Process.CreateKey(cwo.CryptoPassword, asx.Crypto.GetSalt);
-            var pv = new XModel.ProgressViewer();
+            // 복호화 후 원본파일 경로
+            var decryptOriginalFIlePath = cwo.CreateDecryptOriginalFIlePath;
 
-            // 우선 파일 압축을 해제한다
-            FileWork.ZIPExtract(cwo.SourceFilePath, cfn.WorkDirectoryPath, asx.WorkMessage.ZIPExtractFile, pv);
-
-            // 원본파일 해쉬 읽기
-            var originalChecksumChecker = XCrypto.HashWork.ConvertHashText(File.ReadAllBytes(cfn.OriginalChecksumFilePath));
-            // 암호화 된 파일 해쉬 읽기
-            var encryptDataChecksumChecker = XCrypto.HashWork.ConvertHashText(File.ReadAllBytes(cfn.EncryptDataChecksumFilePath));
-            // 복호화 IV 읽기
-            var cryptoIV = File.ReadAllBytes(cfn.CryptoIVFilePath);
-            // 암호화 된 파일 읽기
-            var encryptData = FileWork.GetFileByte(cfn.EncryptDataFilePath, asx.WorkMessage.ReadFile, pv);
-            // 암호화 된 파일 해쉬 만들기
-            var encryptDataChecksum = XCrypto.HashWork.ConvertHashText(XCrypto.HashWork.CreateSHA512(encryptData, asx.WorkMessage.EncryptChecksum, pv));
-
-            // 암호화 된 파일 해쉬 비교
-            if (encryptDataChecksum == encryptDataChecksumChecker)
+            // 원본파일 역시 존재하면 안된다
+            if (File.Exists(decryptOriginalFIlePath) == false)
             {
-                // 파일 복호화
-                var decryptData = XCrypto.AES256X.DecryptNow(cryptoKey, cryptoIV, encryptData, asx.WorkMessage.DecryptFile, pv);
+                // 저장 할 파일들 경로생성
+                var cfn = asx.Crypto.GetCryptoWorkPath;
+                // 복호화 키 생성
+                var cryptoKey = XCrypto.AES256Process.CreateKey(cwo.CryptoPassword, asx.Crypto.GetSalt);
+                var pv = new XModel.ProgressViewer();
+
+                // 우선 파일 압축을 해제한다
+                FileWork.ZIPExtract(cwo.SourceFilePath, cfn.WorkDirectoryPath, asx.WorkMessage.ZIPExtractFile, pv);
+
+                // 원본파일 해쉬 읽기
+                var originalChecksumChecker = XCrypto.HashWork.ConvertHashText(File.ReadAllBytes(cfn.OriginalChecksumFilePath));
+                // 암호화 된 파일 해쉬 읽기
+                var encryptDataChecksumChecker = XCrypto.HashWork.ConvertHashText(File.ReadAllBytes(cfn.EncryptDataChecksumFilePath));
+                // 복호화 IV 읽기
+                var cryptoIV = File.ReadAllBytes(cfn.CryptoIVFilePath);
+                // 암호화 된 파일 읽기
+                var encryptData = FileWork.GetFileByte(cfn.EncryptDataFilePath, asx.WorkMessage.ReadFile, pv);
                 // 암호화 된 파일 해쉬 만들기
-                var decryptDataChecksum = XCrypto.HashWork.ConvertHashText(XCrypto.HashWork.CreateSHA512(decryptData, asx.WorkMessage.DecryptChecksum, pv));
+                var encryptDataChecksum = XCrypto.HashWork.ConvertHashText(XCrypto.HashWork.CreateSHA512(encryptData, asx.WorkMessage.EncryptChecksum, pv));
 
-                if (decryptDataChecksum == originalChecksumChecker)
+                // 암호화 된 파일 해쉬 비교
+                if (encryptDataChecksum == encryptDataChecksumChecker)
                 {
-                    // 원본파일 저장
-                    FileWork.WriteFileByte(decryptData, cwo.CreateDecryptOriginalFIlePath, asx.WorkMessage.SaveDecryptFile, pv);
+                    // 파일 복호화
+                    var decryptData = XCrypto.AES256X.DecryptNow(cryptoKey, cryptoIV, encryptData, asx.WorkMessage.DecryptFile, pv);
+                    // 암호화 된 파일 해쉬 만들기
+                    var decryptDataChecksum = XCrypto.HashWork.ConvertHashText(XCrypto.HashWork.CreateSHA512(decryptData, asx.WorkMessage.DecryptChecksum, pv));
 
-                    // 작업파일 삭제
-                    cfn.DeleteAllFile(cwo.SourceFilePath);
+                    if (decryptDataChecksum == originalChecksumChecker)
+                    {
+                        // 원본파일 저장
+                        FileWork.WriteFileByte(decryptData, decryptOriginalFIlePath, asx.WorkMessage.SaveDecryptFile, pv);
 
-                    // 파일을 복호화 했습니다.
-                    cwms.Success.MessageNow(asx.WorkMessage.DecryptFileDone);
+                        // 작업파일 삭제
+                        cfn.DeleteAllFile(cwo.SourceFilePath);
+
+                        // 파일을 복호화 했습니다.
+                        cwms.Success.MessageNow(asx.WorkMessage.DecryptFileDone);
+                    }
+                    else
+                    {
+                        // 복호화 파일 해쉬가 다릅니다.
+                        cwms.Error.MessageNow(asx.WorkMessage.DifferentDecryptChecksum);
+                    }
                 }
                 else
                 {
-                    // 복호화 파일 해쉬가 다릅니다.
-                    cwms.Error.MessageNow(asx.WorkMessage.DifferentDecryptChecksum);
+                    // 암호화 파일 해쉬가 다릅니다.
+                    cwms.Error.MessageNow(asx.WorkMessage.DifferentEncryptChecksum);
                 }
             }
             else
             {
-                // 암호화 파일 해쉬가 다릅니다.
-                cwms.Error.MessageNow(asx.WorkMessage.DifferentEncryptChecksum);
+                // 같은 이름의 암호화된 파일이 있습니다.
+                // 진행이 중단되었습니다.
+                cwms.Error.MessageNow(asx.WorkMessage.AlreadyExistDecryptFile);
             }
         }
     }
